+++ sbuildr/dependencies/fetchers/git_fetcher.py
@@ -1,4 +1,5 @@
 from sbuildr.dependencies.fetcher import DependencyFetcher
+from sbuildr.logger import G_LOGGER
 from sbuildr.misc import utils
 import subprocess
 import sys
@@ -24,10 +25,16 @@ class GitFetcher(DependencyFetcher):
         super().fetch()
         init_status = subprocess.run(["git", "init"], cwd=self.dest_dir, capture_output=True)
 
+        # Stash any local changes made by external sources
+        G_LOGGER.info(f"Stashing changes in {self.dest_dir}")
+        stash_status = subprocess.run(["git", "stash"], capture_output=False, cwd=self.dest_dir)
+
         checkout = self.commit or self.tag or self.branch
+        G_LOGGER.info(f"Pulling: {self.url} at {checkout} into {self.dest_dir}")
         # # TODO: Error checking here? Pull may fail if this is a local repo.
         pull_status = subprocess.run(["git", "pull", "--force", "--recurse-submodules", "--tags", self.url, checkout], capture_output=False, cwd=self.dest_dir)
 
+        G_LOGGER.info(f"Checking out: {checkout}")
         checkout_status = subprocess.run(["git", "checkout", checkout], capture_output=True, cwd=self.dest_dir)
         if checkout_status.returncode:
             G_LOGGER.critical(f"Failed to checkout {checkout} with:\n{utils.subprocess_output(checkout_status)}")